Add extended Gateway features
